Moved navbar to own library
