+++ apps/user-dashboard/src/components/PageHeader.tsx
@@ -29,7 +29,6 @@ const PageHeaderStyle = styled.header`
   width: calc(100% - 85px);
   z-index: 10;
   height: 85px;
-  background-color: ${(props) => props.theme.base.content};
   justify-content: center;
   h2 {
     font-family: 'Lexend Deca';
